Add Pictures to the About Page
Pictures were added to the about page. The pictures reside in the same
src folder as the rest of the About page, as it is not allowed to import
pictures from the public folder. The pictures are managed by having an
index.js that imports all of them. Then, they can be referred to by
filename in timeline.json. Finally, a README was added to describe how
to add images.